tables

+++ albums_migration.sql
@@ -5,10 +5,11 @@ DROP TABLE IF EXISTS albums;
 CREATE TABLE IF NOT EXISTS albums (
 
   id INT unsigned NOT NULL AUTO_INCREMENT,
-  artist VARCHAR(50),
-  name VARCHAR(50),
+  artist VARCHAR (50),
+  name VARCHAR (50),
   release_date INT(8),
-  sales BIGINT(10),
-  genre VARCHAR(50),
+  sales DECIMAL (3, 2),
+  genre VARCHAR (50),
   PRIMARY KEY (id)
 )
+
